% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_treatment_status}
\alias{plot_treatment_status}
\title{Plot treatment status of individuals over time}
\usage{
plot_treatment_status(
  panel_data,
  Time_name,
  unit_name,
  outcome_name,
  treatment_name,
  final_period,
  initial_period,
  axis.lab.gap = c(0, 0)
)
}
\arguments{
\item{panel_data}{data.frame format. It must be a panel with each row corresponding to a unit at a given date.}

\item{Time_name}{string denoting the column name of the panel corresponding to the date.}

\item{unit_name}{string denoting the column name of the panel corresponding to the individual.}

\item{outcome_name}{string denoting the column name of the panel corresponding to the outcome variable. Rows with missing outcomes will be removed.}

\item{treatment_name}{name of the column of the panel corresponding to the treatment assignment. Rows with missing treatment assignments will be removed.}

\item{final_period}{last period considered}

\item{initial_period}{first period considered}

\item{axis.lab.gap}{gap for x axis}

\item{covariates_names}{vector of strings with column names of the covariates to include in the regression.}
}
\value{
plots
}
\description{
Plot treatment status of individuals over time
}
